+++ src/meson.build
@@ -155,7 +155,6 @@ common_sources = [
   'trx/game/camera/los_camera.c',
   'trx/game/camera/photo_mode.c',
   'trx/game/camera/vars.c',
-  'trx/game/carrier.c',
   'trx/game/catalog/manager.c',
   'trx/game/clock/common.c',
   'trx/game/clock/timer.c',
@@ -281,6 +280,7 @@ common_sources = [
   'trx/game/items/actions/items.c',
   'trx/game/items/actions/lara.c',
   'trx/game/items/anim.c',
+  'trx/game/items/carrier.c',
   'trx/game/items/col.c',
   'trx/game/items/draw.c',
   'trx/game/items/manager.c',
