+++ src/trx/game/items/carrier.c
@@ -1,4 +1,4 @@
-#include <trx/game/carrier.h>
+#include <trx/game/items/carrier.h>
 
 #include <trx/game/game_buf.h>
 #include <trx/game/game_flow.h>
