+++ src/trx/game/objects/creatures/diver.c
@@ -1,6 +1,6 @@
-#include <trx/game/carrier.h>
 #include <trx/game/creature.h>
 #include <trx/game/effects.h>
+#include <trx/game/items/carrier.h>
 #include <trx/game/lara.h>
 #include <trx/game/lara/const.h>
 #include <trx/game/los.h>
