+++ src/trx/game/objects/creatures/mummy.c
@@ -1,7 +1,7 @@
-#include <trx/game/carrier.h>
 #include <trx/game/creature.h>
 #include <trx/game/game.h>
 #include <trx/game/game_buf.h>
+#include <trx/game/items/carrier.h>
 #include <trx/game/lara.h>
 #include <trx/game/math.h>
 #include <trx/game/objects/common.h>
