+++ src/trx/game/objects/creatures/mutant.c
@@ -1,7 +1,7 @@
 #include <trx/game/objects/creatures/mutant.h>
 
-#include <trx/game/carrier.h>
 #include <trx/game/creature.h>
+#include <trx/game/items/carrier.h>
 #include <trx/game/lara.h>
 #include <trx/game/pathing.h>
 #include <trx/game/random.h>
