+++ src/trx/game/objects/creatures/rat.c
@@ -1,5 +1,5 @@
-#include <trx/game/carrier.h>
 #include <trx/game/creature.h>
+#include <trx/game/items/carrier.h>
 #include <trx/game/lara.h>
 #include <trx/game/pathing.h>
 #include <trx/game/random.h>
