+++ src/trx/game/objects/creatures/xian_knight.c
@@ -1,8 +1,8 @@
 #include <trx/debug.h>
-#include <trx/game/carrier.h>
 #include <trx/game/creature.h>
 #include <trx/game/effects.h>
 #include <trx/game/interpolation.h>
+#include <trx/game/items/carrier.h>
 #include <trx/game/lara.h>
 #include <trx/game/objects/common.h>
 #include <trx/game/objects/creatures/xian_common.h>
