+++ src/trx/game/objects/creatures/xian_spearman.c
@@ -1,7 +1,7 @@
 #include <trx/debug.h>
-#include <trx/game/carrier.h>
 #include <trx/game/creature.h>
 #include <trx/game/interpolation.h>
+#include <trx/game/items/carrier.h>
 #include <trx/game/lara.h>
 #include <trx/game/objects.h>
 #include <trx/game/objects/creatures/xian_common.h>
