+++ src/trx/game/savegame/file_read.c
@@ -1,7 +1,6 @@
 #include <trx/config.h>
 #include <trx/debug.h>
 #include <trx/game/camera.h>
-#include <trx/game/carrier.h>
 #include <trx/game/effects.h>
 #include <trx/game/fx/weather.h>
 #include <trx/game/game.h>
@@ -9,6 +8,7 @@
 #include <trx/game/game_flow.h>
 #include <trx/game/gun.h>
 #include <trx/game/inventory.h>
+#include <trx/game/items/carrier.h>
 #include <trx/game/lara.h>
 #include <trx/game/music.h>
 #include <trx/game/objects.h>
