+++ src/trx/game/stats/init.c
@@ -3,11 +3,11 @@
 #include <trx/benchmark.h>
 #include <trx/debug.h>
 #include <trx/filesystem.h>
-#include <trx/game/carrier.h>
 #include <trx/game/creature.h>
 #include <trx/game/game_buf.h>
 #include <trx/game/game_flow.h>
 #include <trx/game/inject.h>
+#include <trx/game/items/carrier.h>
 #include <trx/game/level/format/format.h>
 #include <trx/game/lua.h>
 #include <trx/game/rooms.h>
